Cmdlets refactoring wip
