+++ src/Azos.Sky/Apps/Terminal/Cmdlets/Exec.cs
@@ -8,11 +8,9 @@ using System.Text;
 using Azos.Apps;
 using Azos.Conf;
 using Azos.Glue;
-using Azos.Sky.Workers;
 
-namespace Azos.Sky.Apps.Terminal.Cmdlets
+namespace Azos.Apps.Terminal.Cmdlets
 {
-
   /// <summary>
   /// Execute commands
   /// </summary>
@@ -26,17 +24,20 @@ namespace Azos.Sky.Apps.Terminal.Cmdlets
     {
       var to = m_Args.AttrByName(CONFIG_TO_ATTR).ValueAsString();
       var sb = new StringBuilder();
+
       foreach (var arg in m_Args.Children)
         sb.Append(m_Terminal.Execute(arg));
-      if (to.IsNullOrWhiteSpace())
-        return sb.ToString();
+
+        if (to.IsNullOrWhiteSpace())
+          return sb.ToString();
+
       m_Terminal.Vars[to] = sb.ToString();
       return "OK";
     }
 
     public override string GetHelp()
     {
-      return @"Execute commands";
+      return @"Executes commands";
     }
   }
 }
